perf: Optimize addSingleGroupIntermediateResults for min/max aggregates (#15654)
Summary:
When HashAggregation merges the spilled accumulators into the row container, it
only merges one accumulator of the flat vector at a time. However, the min/max
function allocates an indices vector with a size exceeding 10,000 correspondingly
(this is unnecessary as the spilled vector is flat). As a result, the greater the
amount of spilled data, the higher the cost of allocating the indices vector. This
PR changes the `addSingleGroupIntermediateResults` method to use the index of
the input vector instead. There is a 90% reduction in execution time for one of our
Spark ETL jobs with significant spilling in production.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15654

Reviewed By: kgpai

Differential Revision: D88301999

Pulled By: bikramSingh91

fbshipit-source-id: 50d125a5d78be3a2d091df171d4bd19ff74b34b2