feat: Optimize array_agg to avoid copy (#15715)
Summary:

existing fast path requires a single source and a single contiguous byte range to avoid a copy.
we can apply zero copy to a more sophisticated condition as long as for each group there is a contiguous byte range.

the reason why we see fragments of ranges in one input batch is that in the query plan we have filter and then native join which includes this streamingAgg operator so we fall into the branch where not all rows are selected

Reviewed By: Yuhta, xiaoxmeng

Differential Revision: D88194739