feat: Add listRowsFast in RowContainer (#15698)
Summary:
Adds an optional `useListRowIndex` flag and listRowsFast method in `RowContainer`.
When enabled, RowContainer maintains a row-pointer index (`rowPointers_`) and
uses `listRowsFast()` to iterate rows without scanning allocation ranges or checking
free/probe flags. It is intended to be used in SortBuffer and SortInputSpiller to improve performance.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/15698

Reviewed By: Yuhta

Differential Revision: D88497347

Pulled By: zacw7

fbshipit-source-id: 4cec8eeb6b49fd7a80e97933cb0dd2da038a8a70