feat: Support specififying options for Velox Trace directory (#15735)
Summary:

Introduces `taskTraceDirectoryCreateConfig`, which is passed the trace directory created in `Task.cpp`.

This mirrors the existing `opTraceDirectoryCreateConfig`, which is passed to the trace directory created in `Operator.cpp`.

---

Note: IMO it's not worth having separate configs for these separate use cases. However, the existing config is clearly specific to just `Operator`. I'd rather not make a breaking change by generalizing that. Therefore, I opted to create a new, separate config.

Differential Revision: D87584575