fix: Array concat should require 2 arguments as part of the function signature
Summary:
Today ArrayConcat requires a minimum of 2 arguments when taking a variable number of arrays to concat. This requirement is enforced at runtime when executing the function.

This was inherited from Presto Java, but it leads to inconsistent behavior when expression flattening and constant folding are enabled/disabled.

For example,

concat(concat(array[1, 2], array[3, 4]))

When constant folding is enabled, this becomes

concat(array[1, 2, 3, 4])

This throws an exception because the number of arguments to concat is less than 2.

But when constant folding is disabled and flattening is enabled, this becomes

concat(array[1, 2], array[3, 4])

This is valid and returns array[1, 2, 3, 4]

To fix this, I've updated the signature of concat to take 2 arrays explicitly and then a variable 
number of additional array arguments.  This enforces the minimum 2 arguments at the 
signature level and prevents this inconsistency.

String concat faced the same issue and this approach is consistent with what was done there.
https://github.com/facebookincubator/velox/pull/11076

Differential Revision: D89003931