test: Increase the default bias factor for biased fuzzer runs
Summary:
Today, when we create bias runs we give the functions we're biasing 10 tickets by default,
(other functions have 1).  I've seen this result in a little over 1 in 10 queries using the function
we're biasing for.

Bump up the default to get that ratio higher to improve coverage (we've seen things slip
through despite this).

Differential Revision: D89085415