fix(geo): Handle small multipolygon rings better (#15758)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15758

The ESRI Shapefile serialization uses the orientation of the ring
(clockwise or counterclockwise) to determine whether the ring is the hole of
the current polygon, or the shell of a new polygon.  When the ring is zero
area, the orientation is ill-defined.  In Java the deserializer just guessed,
potentially getting it wrong.  In Velox we raise an error.  Due to float precision,
we had an area threshold of 1e-13. We are seeing polygons in production that have valid
rings that have area 5e-14, so we've decreased the threshold to 1e-15.  The rings that
caused problems before had areas of ~1e-16, so we still have a buffer for numerical
precision.

We had the deserializer raise an internal area, on the theory that we just serialized
it and would have raised a user error for bad rings. However, this theory has been
disproven, possibly due to numerical precision issues. We've converted that to a user
error so that users can TRY and catch those errors.

Reviewed By: Sullivan-Patrick, amitkdutta

Differential Revision: D89077915

fbshipit-source-id: 4c70526e8c179dd1f9adfa548898109c926c0047