test: Increase the default bias factor for biased fuzzer runs (#15762)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15762

Today, when we create bias runs we give the functions we're biasing 10 tickets by default,
(other functions have 1).  I've seen this result in a little over 1 in 10 queries using the function
we're biasing for.

Bump up the default to get that ratio higher to improve coverage (we've seen things slip
through despite this).

Reviewed By: bikramSingh91

Differential Revision: D89085415

fbshipit-source-id: 220f4da6f91cb051cc3c351fe32c71cd5345b4f2