feat(s3fs): Bump AWS_SDK_VERSION from 1.11.321 to 1.11.654 (#14915)
Summary:
Using content-MD5 was an issue in a FIPS enabled environment because MD5 is not allowed to be used and causes SEGVs when used with the FIPS OpenSSL provider. We had this code present, so we explicitly changed the checksum to not use content-MD5.
`request.SetChecksumAlgorithm(Aws::S3::Model::ChecksumAlgorithm::CRC32);`
With the newer version upgrade, we do not see content-md5: header in put request logs so we can remove this.
This is because the default does not send it as we can see from https://docs.aws.amazon.com/sdkref/latest/guide/feature-dataintegrity.html#dataintegrity-sdk-compat

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14915

Reviewed By: kKPulla

Differential Revision: D89202001

Pulled By: kagamiori

fbshipit-source-id: d0cdfdde6cff84c9da2a528c36ed3d102314f0e0