fix(geo): Fix correctness issue with spatial joins on GeometryCollections (#15778)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15778

The envelope for the spatial index was being calculated incorrectly, because
the Shapefile format does not add an envelope to the beginning of a GeometryCollection.
This was causing GeometryCollections to miss hits on the Spatial Index, causing rows
to be dropped.

This uses the newly moved GeometrySerde to get the Envelope in the standard way.
However, this requires us to include GEOS for spatial joins.

Reviewed By: Sullivan-Patrick

Differential Revision: D89239585