Fix string buffer ownership transfer in side table reader
Summary:
During side table merging, if the merged column is a MAP, an output MapVector is initialized and them filled in with values from the base and the side table. When the map contains string types (VARCHAR or VARBINARY), the initialize step also copies over the stringBuffers from the children vectors of the base and the side table so that during the copy phase, it only has the copy the StringView values. This change fixes a bug during the copy phase where if the first row of the first map being copied is empty, then a vector resize() with size 0 is called that clears the stringBuffers. This causes the output map to lose ownership of the string buffers that its values point to. This can result in garbage values being read if those buffers are released.
This change replaces vector resize() with BaseVector::resize(), which only updates nulls buffers and length_, preserving stringBuffers. The merge codepath is responsible for correctly setting vector state as needed.
Discussed and aligned on solution with zzhao0.

Reviewed By: zzhao0

Differential Revision: D89101304