misc: Respect Linux cgroups restrictions when sizing thread pools (#15719)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/15719

This change replaces uses of std::thread::hardware_concurrency with
folly::hardware_concurrency the latter of which always respects
container restrictions while the former does not.  This ensures that
when running inside of a Linux container with restrictions on the
available processors, thread pools and other data structures will not
be over-provisioned.

Reviewed By: kgpai

Differential Revision: D87811255

fbshipit-source-id: 54ad6e6050ae9174acb2ebf3c9ab522035f522a8