fix: Access null value in makeInFilter (#14082)
Summary:
If the value at idx is null, the access to the underlying bit is undefined
behavior, and can lead to a crash, invalid memory access, or other
unpredictable behavior. This PR fixes the access to null value in the
`makeInFilter` by first checking if it is null.

Both Presto and Spark produce the same results for the cases below.
Therefore, during `IN` filter evaluation, NULL values can be ignored.
For `NOT IN` evaluation, if the comparison list contains NULL, the
filter returns always false.

```
select 1 in (2, NULL) -- NULL
select 1 not in (2, NULL) -- NULL

select 2 in (2, NULL) -- true
select 2 not in (2, NULL) -- false

select NULL in (2, NULL) -- NULL
select NULL not in (2, NULL) -- NULL
```

Fixes the core dump issue found in Gluten https://github.com/apache/incubator-gluten/issues/10146.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14082

Reviewed By: kagamiori

Differential Revision: D89406610

Pulled By: kKPulla

fbshipit-source-id: b013a9afbeb40776052a9c102c366eeaa971d602