feat: Add now() and current_timestamp function (#14139)
Summary:
In the past, we were not implementing functions like now() that returned constants in Velox. This was because Presto Java would infer the constant value at the co-ordinator and fold it, so the worker was never passed this function in execution.

However, with recent changes for side-car and worker based constant folding implementation, we need native implementations for functions like now(), so adding to Velox as well.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14139

Reviewed By: peterenescu

Differential Revision: D88676458

Pulled By: kKPulla

fbshipit-source-id: 2aca05a5fac86db2819b5f55ca279b9b34c0a5a6