refactor: Move TableWriteTraits to velox/core (#15835)
Summary:

Moving TableWriteTraits to velox/core, as it is strictly related
to plan metadata, and not to execution. The fact that TableWriteTraits was in
velox/exec forced clients that only wanted to define plans to take a full
dependency on velox/exec.

Leaving an alias in the previous namespace to maintain backward compatibility.

Reviewed By: kagamiori, mbasmanova

Differential Revision: D89592214