feat: Support ArrayOrMapSubscript in Hive connector
Summary:
Add support for the new `ArrayOrMapSubscript` subfield path element.  This new element will only be enabled when plan contains it, so existing use cases are not changed.  Refactor the code to extract the handling of ARRAY and MAP subfields to their own method for readability, no functional change on existing use cases.

Currently the parser only supports `[$]` case, so unit test only has that part.  Other cases will be added when the corresponding parsers are implemented.

Differential Revision: D89565249