refactor(query-ctx): Add builder pattern and documentation
Summary:
Adding a builder pattern for constructing QueryCtx objects. QueryCtx
is part of the external API, and adding members to it over time has cluttered
the create() and constructor methods. It was pretty inconvenient to add and
pass new members to it.

Also adding more documentation and removing unnecessary includes and forward
declarations. No logic changes.

Reviewed By: xiaoxmeng

Differential Revision: D89744645

fbshipit-source-id: a27870a0b17b14ca9420b19727aef4b72187b8f4