feat: Support normalized key encoder in Velox for sort optimizations (#15864)
Summary:

This is to support encode multiple sort columns into a lexicographical comparable string key. This is copy from nimble for index. We might try to leverage this in sort optimization which is a more generic solution compared with normalized key, and to see if it performs better than prefix sort. We might also delete the key encode from nimble later for code consolidation later.

Reviewed By: gggrace14, HuamengJiang, zacw7

Differential Revision: D89846701