fix(ci): Clang-tidy installation
When clang-tidy was moved to the adapters build and fixed the path
(PR #15572)it was missed that it required an install that was done in a previous step.
The PR itself did not run clang-tidy because these was no actual
code change.
One change to the run clang-tidy script is to return 0 if no STDOUT was
generated. Previously, it returned ok only if it matches with a warning
output but without any warning or error it would return fail as well.