refactor(geo): Reduce memcpy in GeometryCollection serialization (#15873)
Summary:

Writing a GeometryCollection requires writing the length of the child geometry
before writing the child geometry. Previously we did this by writing to a
temporary std::string, then writing the length of that string to the output and
then copying the std::string to the output. Recursive GeometryCollections made
this worse because the "output" was actually just the parent std::string.

This commit writes a placeholder size (0), recording the offset. It then writes
the geometry, then writes the resulting length at the offset. This prevents all
the memcpys.

Removing this greatly simplifies our code an allows us to remove
VarbinaryWriter and almost all templating, which futher allows us to move most
implementations into the cpp file.

Differential Revision: D89928986